feat: cross-platform mouse scrolling
